Bump peerDep versions of TSESlint to 7 to match airbnb, avoid conflicts with TSESLint 8
At the current time, attempting to install using the previous peer-dev of v4 throws a version mis-match that npm cannot resolve, as the latest is v8, but AirBnB requires v7 and is NOT compatible with v8. This is because of the ongoing transition to ESLint 9 (and TSESLint 8), which AirBnB does not yet support.

Simply pinning to v7 solves it.